+++ attributes/agent.rb
@@ -1,9 +1,12 @@
-default['logstash']['agent']['version'] = '1.1.13'
+default['logstash']['agent']['version'] = '1.2.2'
+default['logstash']['agent']['home'] = "#{node['logstash']['basedir']}/agent"
 default['logstash']['agent']['log_file'] = '/var/log/logstash/agent.log'
-default['logstash']['agent']['source_url'] = 'https://logstash.objects.dreamhost.com/release/logstash-1.1.13-flatjar.jar'
-default['logstash']['agent']['checksum'] = '5ba0639ff4da064c2a4f6a04bd7006b1997a6573859d3691e210b6855e1e47f1'
+default['logstash']['agent']['source_url'] = 'https://download.elasticsearch.org/logstash/logstash/logstash-1.2.2-flatjar.jar'
+default['logstash']['agent']['checksum'] = '6b0974eed6814f479b68259b690e8c27ecbca2817b708c8ef2a11ce082b1183c'
 default['logstash']['agent']['install_method'] = 'jar' # Either `source` or `jar`
-default['logstash']['agent']['patterns_dir'] = 'agent/etc/patterns'
+default['logstash']['agent']['patterns_dir'] = 'etc/patterns'
+default['logstash']['agent']['config_dir'] = 'etc/conf.d'
+default['logstash']['agent']['config_file'] = 'logstash.conf'
 default['logstash']['agent']['base_config'] = 'agent.conf.erb'
 default['logstash']['agent']['base_config_cookbook'] = 'logstash'
 default['logstash']['agent']['xms'] = '384M'